Move permission resolving to separate service
